requirements.txt
amqp==5.3.1
asgiref==3.10.0
beautifulsoup4==4.14.2
billiard==4.2.2
black==25.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
curl_cffi==0.13.0
Django==5.2.8
django-allauth==65.13.0
django-environ==0.12.0
django-extensions==4.1
django-htmx==1.26.0
django-timezone-field==7.1
flake8==7.3.0
frozendict==2.4.6
idna==3.11
kombu==5.5.4
mccabe==0.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
peewee==3.18.3
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytokens==0.3.0
pytz==2025.2
redis==7.0.1
requests==2.32.5
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
yfinance==0.2.66
